+++ examples/injectable-twitter-server/java/src/test/java/com/twitter/finatra/example/TestQueue.java
@@ -1,5 +1,6 @@
 package com.twitter.finatra.example;
 
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
 import javax.inject.Singleton;
 
@@ -8,15 +9,18 @@ public
 class TestQueue extends Queue {
 
   public final AtomicInteger addCounter;
+  public final CountDownLatch firstWriteLatch;
 
   public TestQueue() {
     super(Integer.MAX_VALUE);
     this.addCounter = new AtomicInteger();
+    this.firstWriteLatch = new CountDownLatch(1);
   }
 
   @Override
   public boolean add(String value) {
     addCounter.incrementAndGet();
+    firstWriteLatch.countDown();
     return super.add(value);
   }
 }